Moved Models to separate Assembly
This has two benefits. First, it allows us to suppress the warnings regarding the missing XmlDocs, which we don't want to require for View Models. Second, it allows these to easily be included or excluded from a project, if it chooses to adopt fully custom View Models.
