+++ Ignia.Topics.ViewModels/ContentListTopicViewModel.cs
