+++ Ignia.Topics.ViewModels/IndexTopicViewModel.cs
