+++ Ignia.Topics.ViewModels/ItemTopicViewModel.cs
