+++ Ignia.Topics.ViewModels/PageGroupTopicViewModel.cs
