+++ Ignia.Topics.ViewModels/PageTopicViewModel.cs
