+++ Ignia.Topics.ViewModels/SectionTopicViewModel.cs
