+++ Ignia.Topics.ViewModels/SlideTopicViewModel.cs
