+++ Ignia.Topics.ViewModels/SlideshowTopicViewModel.cs
