+++ Ignia.Topics.ViewModels/TopicViewModelCollection.cs
