+++ Ignia.Topics.ViewModels/VideoTopicViewModel.cs
