+++ Ignia.Topics.Web.Mvc/TopicViewResult.cs
@@ -9,7 +9,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Web.Mvc;
-using Ignia.Topics.Models;
 
 namespace Ignia.Topics.Web.Mvc {
