+++ Ignia.Topics.sln
@@ -16,6 +16,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ignia.Topics.Web.Mvc", "Ign
 EndProject
 Project("{00D1A9C2-B5F0-4AF3-8072-F6C62B433612}") = "Ignia.Topics.Data.Sql.Database", "Ignia.Topics.Data.Sql.Database\Ignia.Topics.Data.Sql.Database.sqlproj", "{A1D96A34-C6EA-491A-BC60-8BB9E809B005}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ignia.Topics.ViewModels", "Ignia.Topics.ViewModels\Ignia.Topics.ViewModels.csproj", "{E52FC633-B4C5-4A2B-8CAF-30E756D7A6A7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug CC|Any CPU = Debug CC|Any CPU
@@ -68,6 +70,12 @@ Global
 		{A1D96A34-C6EA-491A-BC60-8BB9E809B005}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A1D96A34-C6EA-491A-BC60-8BB9E809B005}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A1D96A34-C6EA-491A-BC60-8BB9E809B005}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{E52FC633-B4C5-4A2B-8CAF-30E756D7A6A7}.Debug CC|Any CPU.ActiveCfg = Debug|Any CPU
+		{E52FC633-B4C5-4A2B-8CAF-30E756D7A6A7}.Debug CC|Any CPU.Build.0 = Debug|Any CPU
+		{E52FC633-B4C5-4A2B-8CAF-30E756D7A6A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E52FC633-B4C5-4A2B-8CAF-30E756D7A6A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E52FC633-B4C5-4A2B-8CAF-30E756D7A6A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E52FC633-B4C5-4A2B-8CAF-30E756D7A6A7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
