+++ Ignia.Topics/Ignia.Topics.csproj
@@ -144,18 +144,6 @@
     <Compile Include="ITopicMappingService.cs" />
     <Compile Include="ITopicRoutingService.cs" />
     <Compile Include="ITopicViewModelCore.cs" />
-    <Compile Include="Models\ContentItemTopicViewModel.cs" />
-    <Compile Include="Models\ContentListTopicViewModel.cs" />
-    <Compile Include="Models\IndexTopicViewModel.cs" />
-    <Compile Include="Models\ItemTopicViewModel.cs" />
-    <Compile Include="Models\PageGroupTopicViewModel.cs" />
-    <Compile Include="Models\PageTopicViewModel.cs" />
-    <Compile Include="Models\SectionTopicViewModel.cs" />
-    <Compile Include="Models\SlideshowTopicViewModel.cs" />
-    <Compile Include="Models\SlideTopicViewItem.cs" />
-    <Compile Include="Models\TopicViewModel.cs" />
-    <Compile Include="Models\TopicViewModelCollection.cs" />
-    <Compile Include="Models\VideoTopicViewModel.cs" />
     <Compile Include="Querying\Topic.cs" />
     <Compile Include="Repositories\DeleteEventArgs.cs" />
     <Compile Include="Repositories\ITopicRepository.cs" />
@@ -173,7 +161,9 @@
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Folder Include="Models\" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\CodeContracts.MSBuild.1.10.1\build\CodeContracts.MSBuild.targets" Condition="Exists('..\packages\CodeContracts.MSBuild.1.10.1\build\CodeContracts.MSBuild.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
